requirements.txt
absl-py==2.0.0
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
beautifulsoup4==4.12.2
black==23.12.1
boto3==1.34.13
botocore==1.34.13
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.9
django-allauth==0.59.0
django-appconf==1.0.6
django-bootstrap-v5==1.0.11
django-compressor==4.4
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-redis==5.4.0
django-sslserver==0.22
django-storages==1.14.2
djangorestframework==3.14.0
exceptiongroup==1.2.0
executing==2.0.1
flatbuffers==23.5.26
gast==0.5.4
google-auth==2.26.1
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.0
gunicorn==21.2.0
h5py==3.10.0
httpie==3.2.2
idna==3.6
importlib-metadata==7.0.1
inotify==0.2.10
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
keras==2.15.0
libclang==16.0.6
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.2.0
multidict==6.0.4
mypy-extensions==1.0.0
nose==1.3.7
numpy==1.26.3
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
protobuf==4.23.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
PyMySQL==1.1.0
pyOpenSSL==23.3.0
PySocks==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
rcssmin==1.1.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.7.0
rjsmin==1.2.1
rsa==4.9
s3transfer==0.10.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
termcolor==2.4.0
tomli==2.0.1
traitlets==5.14.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.26.18
wcwidth==0.2.12
Werkzeug==3.0.1
wrapt==1.14.1
zipp==3.17.0

python-dotenv~=1.0.0